Clean up HumanPlayer class and rename a package
