+++ Sanapuuro/src/main/java/sanapuuro/fileio/LetterReader.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sanapuuro.filereaders;
+package sanapuuro.fileio;
 
 import sanapuuro.utils.LetterFrequencyComparator;
 import java.io.InputStream;
