+++ Sanapuuro/src/main/java/sanapuuro/fileio/WordReader.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sanapuuro.filereaders;
+package sanapuuro.fileio;
 
 import java.io.File;
 import java.io.InputStream;
