--- Tiralabra_maven/src/main/java/com/mycompany/tiralabra_maven/App.java
@@ -1,13 +1,0 @@
-package com.mycompany.tiralabra_maven;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
